requirements.txt
absl-py==2.3.1
accelerate==1.10.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asciitree==0.3.3
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==15.1.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chex==0.1.90
click==8.2.1
cloudpickle==3.1.1
cmake==4.1.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.15
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
deepdiff==8.6.1
deepspeed==0.17.4
diffusers==0.35.1
dill==0.3.8
distrax==0.1.5
distro==1.9.0
dm-tree==0.1.9
draccus==0.11.5
einops==0.8.1
etils==1.13.0
evdev==1.9.2
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.116.1
fasteners==0.20
ffmpeg==1.4
ffmpy==0.6.1
filelock==3.19.1
flash_attn==2.5.7
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gast==0.6.0
gdown==5.2.0
gin-config==0.5.0
git-lfs==1.6
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
gradio==4.44.0
gradio_client==1.3.0
grpcio==1.74.0
gym==0.23.0
gym-notices==0.1.0
gymnasium==0.29.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
inquirerpy==0.3.4
ipykernel==6.30.0
ipython==8.37.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.15.0
kiwisolver==1.4.8
-e git+https://gitclone.com/github.com/huggingface/lerobot@768e36660d1408c71118a2760f831c037fbfa17d#egg=lerobot
libclang==18.1.1
lightning==2.5.2
lightning-utilities==0.15.2
llvmlite==0.44.0
-e git+https://gitclone.com/github.com/simpler-env/ManiSkill2_real2sim@ef7a4d4fdf4b69f2c2154db5b15b9ac8dfe10682#egg=mani_skill2_real2sim
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.4
mergedeep==1.3.4
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numcodecs==0.13.1
numpy==1.26.4
numpydantic==1.6.10
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
oauthlib==3.3.1
# Editable install with no version control (octo==0.0.0)
-e /remote-home/hanmf/SimplerEnv/octo
omegaconf==2.3.0
open_clip_torch==3.1.0
openai==1.108.1
opencv-python==4.9.0.80
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
orbax-checkpoint==0.4.4
orderly-set==5.5.0
orjson==3.11.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.2
parso==0.8.4
peft==0.17.1
pexpect==4.9.0
pfzy==0.3.4
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==10.4.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1653089181607/work
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.1.3
Pygments==2.19.2
pymunk==7.1.0
pynput==1.8.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-multipart==0.0.20
python-xlib==0.33
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
pyzmq==27.0.0
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
rerun-sdk==0.24.1
responses==0.18.0
rich==14.1.0
rsa==4.9.1
rtree==1.4.0
ruckig==0.15.3
ruff==0.12.8
safetensors==0.6.2
sapien==2.2.2
scipy==1.12.0
semantic-version==2.10.0
sentry-sdk==2.37.1
shellingham==1.5.4
# Editable install with no version control (simpler_env==0.0.1)
-e /remote-home/hanmf/SimplerEnv
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.23.0
tensorstore==0.1.76
termcolor==3.1.0
tf-agents==0.19.0
tiktoken==0.6.0
timm==1.0.19
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==1.0.0
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchcodec==0.7.0
torchmetrics==1.8.1
torchvision==0.19.0+cu124
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.0
transforms3d==0.4.2
trimesh==4.7.1
triton==3.0.0
typer==0.16.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wandb==0.21.4
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
wrapt==1.14.1
xxhash==3.5.0
yarl==1.20.1
zarr==2.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
